requirements.txt
appdirs==1.4.3
apturl==0.5.2
attrs==21.4.0
autopage==0.5.0
Babel==2.6.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
boto3==1.24.43
botocore==1.27.43
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cliff==3.10.1
cmd2==2.4.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debtcollector==2.0.0
decorator==4.4.2
deepdiff==5.8.1
defer==1.0.6
deprecation==2.0.7
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
dogpile.cache==0.9.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.7.1
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
iso8601==0.1.12
jmespath==0.9.4
jsonpatch==1.22
jsonpointer==2.0
keyring==18.0.1
keystoneauth1==4.0.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
msgpack==0.6.2
munch==2.3.2
netaddr==0.7.19
netifaces==0.10.4
nftables==0.1
oauthlib==3.1.0
olefile==0.46
openstacksdk==0.61.0
ordered-set==4.1.0
os-client-config==1.31.2
os-service-types==1.7.0
osc-lib==2.5.0
oslo.config==8.0.2
oslo.context==3.0.2
oslo.i18n==4.0.1
oslo.log==4.1.1
oslo.serialization==3.1.1
oslo.utils==4.1.1
packaging==20.3
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
pipenv==2022.7.24
platformdirs==2.5.2
prettytable==0.7.2
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyperclip==1.7.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-cinderclient==7.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-keystoneclient==4.0.0
python-novaclient==17.0.0
python-openstackclient==5.8.0
python-swiftclient==3.9.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
requestsexceptions==1.4.0
rfc3986==1.3.2
roman==2.0.0
s3transfer==0.6.0
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
six==1.14.0
slip==0.6.5
slip.dbus==0.6.5
soupsieve==1.9.5
sshuttle==0.78.5
stevedore==3.5.0
systemd-python==234
testresources==2.0.0
traceback2==1.4.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.16.2
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
xkit==0.0.0
